// src/system/system-extra-charges/enums/system-extra-charges.enum.ts
export enum OperationApplicationEnum {
    sum = 'sum',
    subtraction = 'subtraction',
    division = 'division',
    multiplication = 'multiplication',
}

export enum TypeChargeApplicationEnum {
    percentage = 1,
    quantity = 2,
}

export enum ApplicationFormEnum {
    Manual = 1,
    Automatic = 2,
}