// src/school-colegio-ingles/school-reports/types/school.charge.details.row.type.ts
import { ChargeApplicationEnum, ChargeTypeEnum } from '@munyaal/calculations';

export type SchoolChargeDetailsRow = {
  id: number;
  description: string;
  quantity: number;
  type: ChargeTypeEnum;
  applcation: ChargeApplicationEnum;
  operation: string;
  internalType: string;
  id_detalle: number;
  order: number;
};